+++ app/src/main/java/com/lucas/currencylist/ui/components/CurrencyList.kt
@@ -1,4 +1,4 @@
-package com.lucas.currencylist.ui.screens.currencies
+package com.lucas.currencylist.ui.screens.exchangeList
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
